GERO-295 - refactor - move constants from match.py to constants.py.
