+++ graphkb/util.py
@@ -1,4 +1,3 @@
-import argparse
 import hashlib
 import json
 import logging
@@ -19,23 +18,6 @@ QUERY_CACHE: Dict[Any, Any] = {}
 # https://stackoverflow.com/questions/11029717/how-do-i-disable-log-messages-from-the-requests-library
 
 logger = logging.getLogger('graphkb')
-
-
-class IterableNamespace(argparse.Namespace):
-    def __init__(self, *pos, **kwargs):
-        argparse.Namespace.__init__(self, *pos, **kwargs)
-
-    def keys(self):
-        return self.__dict__.keys()
-
-    def items(self):
-        return self.__dict__.items()
-
-    def values(self):
-        return self.__dict__.values()
-
-    def __getitem__(self, key):
-        return getattr(self, key)
 
 
 def convert_to_rid_list(records: Iterable[Record]) -> List[str]:
