+++ tests/test_genes.py
@@ -5,7 +5,13 @@ import os
 
 import pytest
 
-from graphkb import GraphKBConnection, genes
+from graphkb import GraphKBConnection
+from graphkb.constants import FUSION_NAMES
+from graphkb.genes import (
+    get_genes_from_variant_types,
+    get_oncokb_oncogenes,
+    get_oncokb_tumour_supressors,
+)
 
 CANONICAL_ONCOGENES = ['kras', 'nras', 'alk']
 CANONICAL_TS = ['cdkn2a', 'tp53']
@@ -21,7 +27,7 @@ def conn():
 
 @pytest.mark.parametrize('gene', CANONICAL_ONCOGENES)
 def test_finds_oncogene(conn, gene):
-    result = genes.get_oncokb_oncogenes(conn)
+    result = get_oncokb_oncogenes(conn)
 
     names = {row['name'] for row in result}
     assert gene in names
@@ -29,7 +35,7 @@ def test_finds_oncogene(conn, gene):
 
 @pytest.mark.parametrize('gene', CANONICAL_TS)
 def test_ts_not_oncogene(conn, gene):
-    result = genes.get_oncokb_oncogenes(conn)
+    result = get_oncokb_oncogenes(conn)
 
     names = {row['name'] for row in result}
     assert gene not in names
@@ -37,7 +43,7 @@ def test_ts_not_oncogene(conn, gene):
 
 @pytest.mark.parametrize('gene', CANONICAL_ONCOGENES)
 def test_oncogene_not_ts(conn, gene):
-    result = genes.get_oncokb_tumour_supressors(conn)
+    result = get_oncokb_tumour_supressors(conn)
 
     names = {row['name'] for row in result}
     assert gene not in names
@@ -45,7 +51,7 @@ def test_oncogene_not_ts(conn, gene):
 
 @pytest.mark.parametrize('gene', CANONICAL_TS)
 def test_finds_ts(conn, gene):
-    result = genes.get_oncokb_tumour_supressors(conn)
+    result = get_oncokb_tumour_supressors(conn)
 
     names = {row['name'] for row in result}
     assert gene in names
@@ -53,6 +59,6 @@ def test_finds_ts(conn, gene):
 
 @pytest.mark.parametrize('gene', CANONICAL_FUSION_GENES)
 def test_find_fusion_genes(conn, gene):
-    result = genes.get_genes_from_variant_types(conn, genes.FUSION_NAMES)
+    result = get_genes_from_variant_types(conn, FUSION_NAMES)
     names = {row['name'] for row in result}
     assert gene in names
